requirements.txt
openai>=1.23.0
pillow>=10.3.0
python-dotenv>=1.0.1
tqdm>=4.66.4
requests>=2.32.3
